% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univarkurtosis.R
\name{univar.kurtosis}
\alias{malkaf}
\alias{mardia}
\alias{univar.kurtosis}
\title{Univariate kurtosis indices}
\usage{
univar.krutosis(Omega,alpha,type=c('mardia','malkovich-afifi'))
mardia(Omega,alpha)
malkaf(Omega,alpha)
}
\arguments{
\item{Omega}{a symmetric positive-definite matrix of dimension (d,d)}

\item{alpha}{a numeric vector which regulates the slant of the density}

\item{type}{univariate index chocice, must be 'malkovich-afifi' or 'mardia'. Not required in mardia and malkaf aliases}
}
\value{
The univariate kurtosis index.
}
\description{
Univariate kurtosis indices
}

